Change expansion grammar slightly Fix return value of get_int64 methods()

+++ libs/xml/config.cc
@@ -116,7 +116,7 @@ int Configuration::get_value_int(const string& path, int def)
 // 64-bit integer value fetch
 // Defaults to default value given (or 0) if not present
 // Returns 0 if present but bogus
-int Configuration::get_value_int64(const string& path, uint64_t def)
+uint64_t Configuration::get_value_int64(const string& path, uint64_t def)
 { 
   XPathProcessor xpath(get_root());
   return xpath.get_value_int64(path, def);
